First test edition now works
the script bioconda-recipe-gen.py are now able to build the package
kallisto, detect that it needs the package 'hdf5' on the host to build.
Then run a build again and see it succeed.

To detect that hdf5 is needed on run time, the build.sh needs to be
configured to run some special tests. For more details look at the
run_test.sh file in the kallisto directory in Bioconda-recipe.