Added testing for testing bioconda_recipe_gen
I have added a test file for bioconda_recipe_gen. The file is to be
expanded with more tests, as we generalize bioconda_recipe_gen.

Right now there is a single test, for a ned method added to
bioconda_recipe_gen for the purpose of showing how to import the module
for testing.