Alpine build: source is copied on docker build
When building in Alpine, the source will now first get downloaded an
unpacked.
Then the image is build with `docker build`, in this stage the
source is copied into the image.
This solution will decouple the source directory from the image and
the tempdir can now be cleaned without root access.