Resolved #16
There is added a MANIFEST.in file, which defines resources which should
be packed with the code.
The setup.py is configured to look for the MANIFEST.in file.
Bioconda_recipe_gen.py will now aquire the templates with the
pkg_resources library instead of just copying the template files.