The script can now run a package test, the resolves #11
The function alpine_run_test is now in place. It will take a package, a
test and the build dependecies for the package.
Then it will build the package and try to run the test. It will then try
to make a list of runtime dependencies.
There has also been written a test, which does this process with the
kallisto package.