Impl multi build strategy. But not working fully.
The code still builds the packages it builded before.
(Tested on kallisto, newick-utils and qfilt.) It also
generates the exact same meta.yaml and build.sh as was used
24.9.2019 where our program builded netreg. But for some
reason it doesn't build now.

This commit also includes the new Filesystem class, which is
added to each the buildscript instance as a property called
'filesystem'.